+++ packages/libs/react-ui/src/components/Form/InputWrapper/InputWrapper.tsx
@@ -2,17 +2,18 @@ import type { IInputProps } from '@components/Form';
 import type { FC, FunctionComponentElement } from 'react';
 import React from 'react';
 import type { vars } from 'src/styles';
+import type { FormFieldStatus } from '../Form.css';
 import type { IInputHeaderProps } from './InputHeader/InputHeader';
 import { InputHeader } from './InputHeader/InputHeader';
 import { InputHelper } from './InputHelper/InputHelper';
-import type { Status } from './InputWrapper.css';
+import { InputWrapperContext } from './InputWrapper.context';
 import { statusVariant } from './InputWrapper.css';
 
 export interface IInputWrapperProps extends Omit<IInputHeaderProps, 'label'> {
   children:
     | FunctionComponentElement<IInputProps>
     | FunctionComponentElement<IInputProps>[];
-  status?: Status;
+  status?: FormFieldStatus;
   disabled?: boolean;
   helperText?: string;
   label?: string;
@@ -33,24 +34,14 @@ export const InputWrapper: FC<IInputWrapperProps> = ({
   const statusVal = disabled === true ? 'disabled' : status;
 
   return (
-    <div className={statusVal ? statusVariant[statusVal] : undefined}>
-      {label !== undefined && (
-        <InputHeader htmlFor={htmlFor} label={label} tag={tag} info={info} />
-      )}
-      <div className="inputGroup">
-        {leadingTextWidth
-          ? React.Children.map(children, (child) => {
-              if (!React.isValidElement(child)) return null;
-              const props = {
-                ...child.props,
-                leadingTextWidth,
-              };
-
-              return React.cloneElement(child, props);
-            })
-          : children}
+    <InputWrapperContext.Provider value={{ status, leadingTextWidth }}>
+      <div className={statusVal ? statusVariant[statusVal] : undefined}>
+        {label !== undefined && (
+          <InputHeader htmlFor={htmlFor} label={label} tag={tag} info={info} />
+        )}
+        <div className="inputGroup">{children}</div>
+        {Boolean(helperText) && <InputHelper>{helperText}</InputHelper>}
       </div>
-      {Boolean(helperText) && <InputHelper>{helperText}</InputHelper>}
-    </div>
+    </InputWrapperContext.Provider>
   );
 };