+++ packages/libs/react-ui/src/components/Layout/Box/Box.stories.tsx
@@ -3,7 +3,7 @@ import { Box } from '@components/Layout/Box';
 import type { Meta, StoryObj } from '@storybook/react';
 import { vars } from '@theme/vars.css';
 import React from 'react';
-import { containerClass, itemClass } from '../stories.css';
+import { componentClass, containerClass, itemClass } from '../stories.css';
 
 const spaceOptions: (keyof typeof vars.sizes | undefined)[] = [
   undefined,
@@ -17,6 +17,7 @@ const dimensionOptions: string[] = ['100%', 'min-content', 'max-content'];
 
 const meta: Meta<IBoxProps> = {
   title: 'Layout/Box',
+  component: Box,
   parameters: {
     docs: {
       description: {
@@ -27,6 +28,13 @@ const meta: Meta<IBoxProps> = {
     },
   },
   argTypes: {
+    overflow: {
+      options: ['hidden', 'visible', 'scroll', 'auto'],
+      control: {
+        type: 'select',
+      },
+      description: 'Overflow css property.',
+    },
     width: {
       options: [...spaceOptions, ...dimensionOptions, ...contentWidthOptions],
       control: {
@@ -218,7 +226,7 @@ export const Primary: Story = {
     marginBottom,
     marginLeft,
     marginRight,
-    padding = '$6',
+    padding,
     paddingX,
     paddingY,
     paddingTop,
@@ -231,6 +239,7 @@ export const Primary: Story = {
     height,
     minHeight,
     maxHeight,
+    overflow,
   }) => (
     <div className={containerClass}>
       <Box
@@ -254,9 +263,10 @@ export const Primary: Story = {
         height={height}
         minHeight={minHeight}
         maxHeight={maxHeight}
-        className={itemClass}
+        overflow={overflow}
+        className={componentClass}
       >
-        Box Content
+        <div className={itemClass}>Box Content</div>
       </Box>
     </div>
   ),