+++ packages/libs/react-ui/src/index.ts
@@ -3,6 +3,7 @@ import './styles/global.css';
 KodeMono();
 
 export type {
+  FormFieldStatus,
   IAccordionProps,
   IAccordionRootProps,
   IAccordionSectionProps,
@@ -62,7 +63,6 @@ export type {
   ITrProps,
   ITrackerCardProps,
   ITreeProps,
-  InputWrapperStatus,
   MaskOptions,
 } from './components';