+++ packages/tools/cookbook/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @kadena/cookbook
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [445fb2c7d]
+  - @kadena/client@1.6.1
+
 ## 0.0.5
 
 ### Patch Changes